requirements-local.txt
# only install locally since GAE has its own version
MySQL-python==1.2.5
djangorestframework==2.3.12
